+++ src/main/java/org/clueweb/clueweb12/app/BuildPForDocVectors.java
@@ -18,7 +18,7 @@ package org.clueweb.clueweb12.app;
 
 import java.io.IOException;
 import java.util.Arrays;
-import java.util.Map;
+import java.util.List;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
@@ -34,7 +34,6 @@ import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.Mapper;
-import org.apache.hadoop.mapreduce.Reducer;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
 import org.apache.hadoop.mapreduce.lib.output.SequenceFileOutputFormat;
@@ -46,30 +45,34 @@ import org.apache.lucene.analysis.standard.StandardAnalyzer;
 import org.apache.lucene.util.Version;
 import org.clueweb.clueweb12.mapreduce.ClueWarcInputFormat;
 import org.clueweb.data.ClueWarcRecord;
+import org.clueweb.data.PForDocVector;
+import org.clueweb.dictionary.DefaultFrequencySortedDictionary;
 import org.jsoup.Jsoup;
 
-import tl.lin.data.pair.PairOfIntLong;
+import tl.lin.data.array.IntArrayWritable;
 import tl.lin.lucene.AnalyzerUtils;
 
-import com.google.common.collect.Maps;
+public class BuildPForDocVectors extends Configured implements Tool {
+  private static final Logger LOG = Logger.getLogger(BuildPForDocVectors.class);
 
-public class ComputeTermStatistics extends Configured implements Tool {
-  private static final Logger LOG = Logger.getLogger(ComputeTermStatistics.class);
-
-  private static enum Records { TOTAL, PAGES, ERRORS, SKIPPED };
+  private static enum Records { TOTAL, PAGES, ERRORS, TOO_LONG };
 
   private static final Analyzer ANALYZER = new StandardAnalyzer(Version.LUCENE_43);
 
-  private static final String HADOOP_DF_MIN_OPTION = "df.min";
-  private static final String HADOOP_DF_MAX_OPTION = "df.max";
-  
-  private static final int MAX_TOKEN_LENGTH = 64;       // Throw away tokens longer than this.
-  private static final int MIN_DF_DEFAULT = 100;        // Throw away terms with df less than this.
   private static final int MAX_DOC_LENGTH = 512 * 1024; // Skip document if long than this.
 
-  private static class MyMapper extends Mapper<LongWritable, ClueWarcRecord, Text, PairOfIntLong> {
-    private static final Text term = new Text();
-    private static final PairOfIntLong pair = new PairOfIntLong();
+  private static class MyMapper extends Mapper<LongWritable, ClueWarcRecord, Text, IntArrayWritable> {
+    private static final Text DOCID = new Text();
+    private static final IntArrayWritable DOC = new IntArrayWritable();
+
+    private DefaultFrequencySortedDictionary dictionary;
+
+    @Override
+    public void setup(Context context) throws IOException {
+      FileSystem fs = FileSystem.get(context.getConfiguration());
+      String path = context.getConfiguration().get(DICTIONARY_OPTION);
+      dictionary = new DefaultFrequencySortedDictionary(path, fs);
+    }
 
     @Override
     public void map(LongWritable key, ClueWarcRecord doc, Context context)
@@ -79,6 +82,8 @@ public class ComputeTermStatistics extends Configured implements Tool {
 
       String docid = doc.getHeaderMetadataItem("WARC-TREC-ID");
       if (docid != null) {
+        DOCID.set(docid);
+
         context.getCounter(Records.PAGES).increment(1);
         try {
           String content = doc.getContent();
@@ -89,87 +94,44 @@ public class ComputeTermStatistics extends Configured implements Tool {
           //    http://stackoverflow.com/questions/2275443/how-to-timeout-a-thread
           if ( content.length() > MAX_DOC_LENGTH ) {
             LOG.info("Skipping " + docid + " due to excessive length: " + content.length());
-            context.getCounter(Records.SKIPPED).increment(1);
+            context.getCounter(Records.TOO_LONG).increment(1);
+            PForDocVector.toIntArrayWritable(DOC, new int[] {}, 0);
+            context.write(DOCID, DOC);
             return;
           }
 
           String cleaned = Jsoup.parse(content).text();
-          Map<String, Integer> map = Maps.newHashMap();
-          for (String term : AnalyzerUtils.parse(ANALYZER, cleaned)) {
-            if (term.length() > MAX_TOKEN_LENGTH) {
-              continue;
-            }
+          List<String> tokens = AnalyzerUtils.parse(ANALYZER, cleaned);
 
-            if (map.containsKey(term)) {
-              map.put(term, map.get(term) + 1);
-            } else {
-              map.put(term, 1);
+          int len = 0;
+          int[] termids = new int[tokens.size()];
+          for (String token : tokens) {
+            int id = dictionary.getId(token);
+            if (id != -1) {
+              termids[len] = id;
+              len++;
             }
           }
 
-          for (Map.Entry<String, Integer> entry : map.entrySet()) {
-            term.set(entry.getKey());
-            pair.set(1, entry.getValue());
-            context.write(term, pair);
-          }
-        } catch (Exception e) {
-          // If Jsoup throws any exceptions, catch and move on.
+          PForDocVector.toIntArrayWritable(DOC, termids, len);
+          context.write(DOCID, DOC);
+        }
+        catch (Exception e) {
+          // If Jsoup throws any exceptions, catch and move on, but emit empty doc.
           LOG.info("Error caught processing " + docid);
+          DOC.setArray(new int[] {});  // Clean up possible corrupted data
           context.getCounter(Records.ERRORS).increment(1);
+          PForDocVector.toIntArrayWritable(DOC, new int[] {}, 0);
+          context.write(DOCID, DOC);
         }
       }
     }
   }
 
-  private static class MyCombiner extends Reducer<Text, PairOfIntLong, Text, PairOfIntLong> {
-    private static final PairOfIntLong output = new PairOfIntLong();
-
-    @Override
-    public void reduce(Text key, Iterable<PairOfIntLong> values, Context context)
-    throws IOException, InterruptedException {
-      int df = 0;
-      long cf = 0;
-      for (PairOfIntLong pair : values) {
-        df += pair.getLeftElement();
-        cf += pair.getRightElement();
-      }
-
-      output.set(df, cf);
-      context.write(key, output);
-    }
-  }
-
-  private static class MyReducer extends Reducer<Text, PairOfIntLong, Text, PairOfIntLong> {
-    private static final PairOfIntLong output = new PairOfIntLong();
-    private int dfMin, dfMax;
-
-    @Override
-    public void setup(Reducer<Text, PairOfIntLong, Text, PairOfIntLong>.Context context) {
-      dfMin = context.getConfiguration().getInt(HADOOP_DF_MIN_OPTION, MIN_DF_DEFAULT);
-      dfMax = context.getConfiguration().getInt(HADOOP_DF_MAX_OPTION, Integer.MAX_VALUE);
-      LOG.info("dfMin = " + dfMin);
-    }
-
-    @Override
-    public void reduce(Text key, Iterable<PairOfIntLong> values, Context context)
-    throws IOException, InterruptedException {
-      int df = 0;
-      long cf = 0;
-      for (PairOfIntLong pair : values) {
-        df += pair.getLeftElement();
-        cf += pair.getRightElement();
-      }
-      if (df < dfMin || df > dfMax) {
-        return;
-      }
-      output.set(df, cf);
-      context.write(key, output);
-    }
-  }
-
   public static final String INPUT_OPTION = "input";
   public static final String OUTPUT_OPTION = "output";
-  public static final String DF_MIN_OPTION = "dfMin";
+  public static final String DICTIONARY_OPTION = "dictionary";
+  public static final String REDUCERS_OPTION = "reducers";
 
   /**
    * Runs this tool.
@@ -182,8 +144,10 @@ public class ComputeTermStatistics extends Configured implements Tool {
         .withDescription("input path").create(INPUT_OPTION));
     options.addOption(OptionBuilder.withArgName("path").hasArg()
         .withDescription("output path").create(OUTPUT_OPTION));
+    options.addOption(OptionBuilder.withArgName("path").hasArg()
+        .withDescription("dictionary").create(DICTIONARY_OPTION));
     options.addOption(OptionBuilder.withArgName("num").hasArg()
-        .withDescription("minimum df").create(DF_MIN_OPTION));
+        .withDescription("number of reducers").create(REDUCERS_OPTION));
 
     CommandLine cmdline;
     CommandLineParser parser = new GnuParser();
@@ -197,7 +161,8 @@ public class ComputeTermStatistics extends Configured implements Tool {
       return -1;
     }
 
-    if (!cmdline.hasOption(INPUT_OPTION) || !cmdline.hasOption(OUTPUT_OPTION)) {
+    if (!cmdline.hasOption(INPUT_OPTION) || !cmdline.hasOption(OUTPUT_OPTION) ||
+        !cmdline.hasOption(DICTIONARY_OPTION)) {
       HelpFormatter formatter = new HelpFormatter();
       formatter.printHelp(this.getClass().getName(), options);
       ToolRunner.printGenericCommandUsage(System.out);
@@ -206,36 +171,38 @@ public class ComputeTermStatistics extends Configured implements Tool {
 
     String input = cmdline.getOptionValue(INPUT_OPTION);
     String output = cmdline.getOptionValue(OUTPUT_OPTION);
+    String dictionary = cmdline.getOptionValue(DICTIONARY_OPTION);
 
-    LOG.info("Tool name: " + ComputeTermStatistics.class.getSimpleName());
+    Job job = new Job(getConf(), BuildPForDocVectors.class.getSimpleName() + ":" + input);
+    job.setJarByClass(BuildPForDocVectors.class);
+
+    LOG.info("Tool name: " + BuildPForDocVectors.class.getSimpleName());
     LOG.info(" - input: " + input);
     LOG.info(" - output: " + output);
+    LOG.info(" - dictionary: " + dictionary);
 
-    Job job = new Job(getConf(), ComputeTermStatistics.class.getSimpleName() + ":" + input);
-    job.setJarByClass(ComputeTermStatistics.class);
-
-    job.setNumReduceTasks(100);
-
-    if (cmdline.hasOption(DF_MIN_OPTION)) {
-      int dfMin = Integer.parseInt(cmdline.getOptionValue(DF_MIN_OPTION));
-      LOG.info(" - dfMin: " + dfMin);
-      job.getConfiguration().setInt(HADOOP_DF_MIN_OPTION, dfMin);
+    if (cmdline.hasOption(REDUCERS_OPTION)) {
+      int numReducers = Integer.parseInt(cmdline.getOptionValue(REDUCERS_OPTION));
+      LOG.info(" - reducers: " + numReducers);
+      job.setNumReduceTasks(numReducers);
+    } else {
+      job.setNumReduceTasks(0);
     }
 
     FileInputFormat.setInputPaths(job, input);
     FileOutputFormat.setOutputPath(job, new Path(output));
 
+    job.getConfiguration().set(DICTIONARY_OPTION, dictionary);
+
     job.setInputFormatClass(ClueWarcInputFormat.class);
     job.setOutputFormatClass(SequenceFileOutputFormat.class);
 
     job.setMapOutputKeyClass(Text.class);
-    job.setMapOutputValueClass(PairOfIntLong.class);
+    job.setMapOutputValueClass(IntArrayWritable.class);
     job.setOutputKeyClass(Text.class);
-    job.setOutputValueClass(PairOfIntLong.class);
+    job.setOutputValueClass(IntArrayWritable.class);
 
     job.setMapperClass(MyMapper.class);
-    job.setCombinerClass(MyCombiner.class);
-    job.setReducerClass(MyReducer.class);
 
     FileSystem.get(getConf()).delete(new Path(output), true);
 
@@ -250,8 +217,8 @@ public class ComputeTermStatistics extends Configured implements Tool {
    * Dispatches command-line arguments to the tool via the <code>ToolRunner</code>.
    */
   public static void main(String[] args) throws Exception {
-    LOG.info("Running " + ComputeTermStatistics.class.getCanonicalName() + " with args "
+    LOG.info("Running " + BuildPForDocVectors.class.getCanonicalName() + " with args "
         + Arrays.toString(args));
-    ToolRunner.run(new ComputeTermStatistics(), args);
+    ToolRunner.run(new BuildPForDocVectors(), args);
   }
 }