+++ src/main/java/org/clueweb/clueweb12/app/CountClueWarcRecordsNew.java
@@ -1,3 +1,19 @@
+/*
+ * ClueWeb Tools: Hadoop tools for manipulating ClueWeb collections
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you
+ * may not use this file except in compliance with the License. You may
+ * obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
 package org.clueweb.clueweb12.app;
 
 import java.io.IOException;
@@ -21,8 +37,8 @@ import org.apache.hadoop.mapreduce.lib.output.NullOutputFormat;
 import org.apache.hadoop.util.Tool;
 import org.apache.hadoop.util.ToolRunner;
 import org.apache.log4j.Logger;
-import org.clueweb.clueweb12.data.ClueWarcRecord;
 import org.clueweb.clueweb12.mapreduce.ClueWarcInputFormat;
+import org.clueweb.data.ClueWarcRecord;
 
 public class CountClueWarcRecordsNew extends Configured implements Tool {
   private static final Logger LOG = Logger.getLogger(CountClueWarcRecordsNew.class);