+++ src/main/java/org/clueweb/clueweb12/app/DumpClueWarcRecordsToPlainText.java
@@ -1,3 +1,19 @@
+/*
+ * ClueWeb Tools: Hadoop tools for manipulating ClueWeb collections
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you
+ * may not use this file except in compliance with the License. You may
+ * obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
 package org.clueweb.clueweb12.app;
 
 import java.io.IOException;
@@ -31,10 +47,11 @@ import org.apache.log4j.Logger;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
 import org.apache.lucene.util.Version;
-import org.clueweb.clueweb12.data.ClueWarcRecord;
-import org.clueweb.clueweb12.lucene.AnalyzerUtil;
 import org.clueweb.clueweb12.mapred.ClueWarcInputFormat;
+import org.clueweb.data.ClueWarcRecord;
 import org.jsoup.Jsoup;
+
+import tl.lin.lucene.AnalyzerUtils;
 
 import com.google.common.base.Joiner;
 
@@ -62,7 +79,7 @@ public class DumpClueWarcRecordsToPlainText extends Configured implements Tool {
         try {
           KEY.set(docid);
           String cleaned = Jsoup.parse(doc.getContent()).text().replaceAll("[\\r\\n]+", " ");
-          cleaned = JOINER.join(AnalyzerUtil.parse(ANALYZER, cleaned));
+          cleaned = JOINER.join(AnalyzerUtils.parse(ANALYZER, cleaned));
           VALUE.set(cleaned);
           output.collect(KEY, VALUE);
         } catch (Exception e) {