+++ src/main/java/org/clueweb/clueweb12/dictionary/FrequencySortedDictionary.java
@@ -1,9 +1,0 @@
-package org.clueweb.clueweb12.dictionary;
-
-/**
- * A frequency-sorted dictionary. That is, smaller term ids are assigned to more
- * frequently occurring terms.
- *
- * @author Jimmy Lin
- */
-public interface FrequencySortedDictionary extends Dictionary {}