+++ src/main/java/org/clueweb/clueweb12/dictionary/LexicographicallySortedDictionary.java
@@ -1,9 +1,0 @@
-package org.clueweb.clueweb12.dictionary;
-
-/**
- * A lexicographically-sorted dictionary. That is, smaller term ids correspond to terms
- * that are sorted lexicographically earlier.
- *
- * @author Jimmy Lin
- */
-public interface LexicographicallySortedDictionary extends Dictionary {}