+++ src/main/java/org/clueweb/dictionary/DefaultFrequencySortedDictionary.java
@@ -1,4 +1,20 @@
-package org.clueweb.clueweb12.dictionary;
+/*
+ * ClueWeb Tools: Hadoop tools for manipulating ClueWeb collections
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you
+ * may not use this file except in compliance with the License. You may
+ * obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+package org.clueweb.dictionary;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -11,7 +27,7 @@ import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.clueweb.clueweb12.app.BuildDictionary;
-import org.clueweb.clueweb12.data.TermStatistics;
+import org.clueweb.data.TermStatistics;
 
 /**
  * An implementation of {@link FrequencySortedDictionary}. Term ids start at 1, which corresponds to
@@ -130,6 +146,7 @@ public class DefaultFrequencySortedDictionary implements FrequencySortedDictiona
     TermStatistics stats = new TermStatistics(new Path(path), fs);
     out.println("max df = " + stats.getMaxDf() + ", termid " + stats.getMaxDfTerm());
     out.println("max cf = " + stats.getMaxCf() + ", termid " + stats.getMaxCfTerm());
+    out.println("collection size = " + stats.getCollectionSize());
     out.println("");
 
     out.println(" \"term word\" to lookup termid; \"termid 234\" to lookup term");